moved some gcp utils and types into shell
