+++ pkg/gke/components/__tests__/Import.test.ts
@@ -28,7 +28,7 @@ const requiredSetup = () => {
   };
 };
 
-jest.mock('@pkg/gke/util/gcp');
+jest.mock('@shell/components/google/util/gcp');
 jest.mock('lodash/debounce', () => jest.fn((fn) => fn));
 
 describe('gke import', () => {
