+++ pkg/gke/components/__tests__/Networking.test.ts
@@ -37,7 +37,7 @@ const requiredSetup = () => {
   };
 };
 
-jest.mock('@pkg/gke/util/gcp');
+jest.mock('@shell/components/google/util/gcp');
 jest.mock('lodash/debounce', () => jest.fn((fn) => fn));
 
 describe('gke Networking', () => {
