+++ shell/components/google/types/gcp.d.ts
