+++ shell/components/google/util/__mocks__/gcp.ts
