+++ shell/machine-config/components/GCEImage.vue
@@ -4,7 +4,7 @@ import { LabeledInput } from '@components/Form/LabeledInput';
 import { Checkbox } from '@components/Form/Checkbox';
 import FormValidation from '@shell/mixins/form-validation';
 import { _CREATE } from '@shell/config/query-params';
-import { getGKEImageFamilies, getGKEFamiliesFromProject } from '@pkg/gke/util/gcp';
+import { getGKEImageFamilies, getGKEFamiliesFromProject } from '@shell/components/google/util/gcp';
 import debounce from 'lodash/debounce';
 import { mapGetters } from 'vuex';
